Add method to calculate if a player's pick is unique and display it
My wife wrote all this code - you can tell now that we are
embedding SQL statements and not using SQLAlachemy.  And I'm so
grateful because it probably would have taken me days if not weeks
to do what she just did in a few hours.  And a new column displays
when a player views their picks on their account page showing if
a pick is unique.