Update Pendulum use to Pendulum 2.x
Use MSF to get the date of the first game and remove data input from
admin.  Use MSF data and Pendulum to add that data.  Change db to
make it a Datetime field to store that info and calculate that the
time of the first game is correct (PM, not AM)